+++ Ticket monitor/Ticket monitor/Entities/SLFService.h
@@ -11,7 +11,7 @@
 
 @interface SLFService : NSObject <NSCoding, NSCopying>
 
-@property (nonatomic, assign) double ID;
+@property (nonatomic, assign) int ID;
 @property (nonatomic, strong) NSString *name;
 
 + (instancetype)modelObjectWithDictionary:(NSDictionary *)dict;
