+++ Ticket monitor/Ticket monitor/Entities/SLFSubscription.h
@@ -12,7 +12,7 @@
 @interface SLFSubscription : NSObject <NSCoding, NSCopying>
 
 @property (nonatomic, assign) BOOL active;
-@property (nonatomic, assign) double ruleTypeID;
+@property (nonatomic, assign) int ruleTypeID;
 @property (nonatomic, strong) NSString *subscriptionGroupID;
 @property (nonatomic, strong) NSString *lastCheckPoint;
 @property (nonatomic, strong) NSString *valueDisplayText;
